refactor: move components around
