+++ pages/posts.tsx
@@ -1,7 +1,7 @@
 import { InferGetStaticPropsType } from 'next';
 import Link from 'next/link';
 import { getAllPosts } from '../api/posts';
-import Wrapper from '../components/layout/Wrapper';
+import Wrapper from '../components/Wrapper';
 import PageHeader from '../components/PageHeader';
 import PostDate from '../components/PostDate';
